Android 15.0.0 Release 6 (AP4A.241205.013)

* tag 'android-15.0.0_r6': (339 commits)
  Remove the shared dex data handling from oat_writer.
  Fix record trace entry points to correctly update entries
  Update volatile variables to C++20
  Store reference to the target class in MH object.
  Try to fix 2276-const-method-type-gc-cleanup.
  More suspend timeout improvements
  Revert "Do not unmap twice a mapping."
  LUCI: Revert 'Checkout full android tree for qemu-armv8'
  Check receiver for null in invokeExact intrinsic.
  Fix incorrect assertion checking default values of system properties
  Offload `PreRebootDexoptJob.onUpdateReady` to a separate thread.
  Report dex2oat metrics to StatsD when dexopting
  Do not unmap twice a mapping.
  Add missing new line before dumping image spaces.
  madvise moving space after uffd registration
  Lower annotation log to WARNING
  Add missing return statement in UnstartedClassGetEnclosingClass
  Print the full relative address in Arm64
  Improve the GC strategy for the gtest of the class verification fuzzer's corpus
  Check that method is defined in the class in...
  ...

+++ runtime/runtime_intrinsics.cc
@@ -63,10 +63,10 @@ bool InitializeIntrinsic(Thread* self,
 
   CHECK_EQ(method->GetInvokeType(), invoke_type);
   if (method->IsIntrinsic()) {
-    CHECK_EQ(method->GetIntrinsic(), static_cast<uint32_t>(intrinsic));
+    CHECK_EQ(method->GetIntrinsic(), intrinsic);
     return true;
   } else {
-    method->SetIntrinsic(static_cast<uint32_t>(intrinsic));
+    method->SetIntrinsic(intrinsic);
     return false;
   }
 }
@@ -83,7 +83,7 @@ bool IsIntrinsicInitialized(Thread* self,
 
   CHECK_EQ(method->GetInvokeType(), invoke_type);
   if (method->IsIntrinsic()) {
-    CHECK_EQ(method->GetIntrinsic(), static_cast<uint32_t>(intrinsic));
+    CHECK_EQ(method->GetIntrinsic(), intrinsic);
     return true;
   } else {
     return false;
